+++ portlets/kaleo-forms-portlet/docroot/WEB-INF/service/com/liferay/portal/workflow/kaleo/forms/service/KaleoProcessLocalServiceWrapper.java
@@ -51,7 +51,7 @@ public class KaleoProcessLocalServiceWrapper implements KaleoProcessLocalService
 		java.util.Map<java.util.Locale, java.lang.String> nameMap,
 		java.util.Map<java.util.Locale, java.lang.String> descriptionMap,
 		long ddmTemplateId, java.lang.String workflowDefinitionName,
-		long workflowDefinitionVersion,
+		int workflowDefinitionVersion,
 		com.liferay.portal.workflow.kaleo.forms.util.TaskFormPairs taskFormPairs,
 		com.liferay.portal.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
@@ -327,7 +327,7 @@ public class KaleoProcessLocalServiceWrapper implements KaleoProcessLocalService
 		java.util.Map<java.util.Locale, java.lang.String> nameMap,
 		java.util.Map<java.util.Locale, java.lang.String> descriptionMap,
 		long ddmTemplateId, java.lang.String workflowDefinitionName,
-		long workflowDefinitionVersion,
+		int workflowDefinitionVersion,
 		com.liferay.portal.workflow.kaleo.forms.util.TaskFormPairs taskFormPairs,
 		com.liferay.portal.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
