+++ portlets/kaleo-forms-portlet/docroot/WEB-INF/service/com/liferay/portal/workflow/kaleo/forms/service/KaleoProcessServiceClp.java
@@ -31,7 +31,7 @@ public class KaleoProcessServiceClp implements KaleoProcessService {
 
 		_methodParameterTypes0 = new String[] {
 				"long", "long", "java.util.Map", "java.util.Map", "long",
-				"java.lang.String", "long",
+				"java.lang.String", "int",
 				"com.liferay.portal.workflow.kaleo.forms.util.TaskFormPairs",
 				"com.liferay.portal.service.ServiceContext"
 			};
@@ -67,7 +67,7 @@ public class KaleoProcessServiceClp implements KaleoProcessService {
 
 		_methodParameterTypes8 = new String[] {
 				"long", "long", "java.util.Map", "java.util.Map", "long",
-				"java.lang.String", "long",
+				"java.lang.String", "int",
 				"com.liferay.portal.workflow.kaleo.forms.util.TaskFormPairs",
 				"com.liferay.portal.service.ServiceContext"
 			};
@@ -79,7 +79,7 @@ public class KaleoProcessServiceClp implements KaleoProcessService {
 		java.util.Map<java.util.Locale, java.lang.String> nameMap,
 		java.util.Map<java.util.Locale, java.lang.String> descriptionMap,
 		long ddmTemplateId, java.lang.String workflowDefinitionName,
-		long workflowDefinitionVersion,
+		int workflowDefinitionVersion,
 		com.liferay.portal.workflow.kaleo.forms.util.TaskFormPairs taskFormPairs,
 		com.liferay.portal.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
@@ -298,7 +298,7 @@ public class KaleoProcessServiceClp implements KaleoProcessService {
 		java.util.Map<java.util.Locale, java.lang.String> nameMap,
 		java.util.Map<java.util.Locale, java.lang.String> descriptionMap,
 		long ddmTemplateId, java.lang.String workflowDefinitionName,
-		long workflowDefinitionVersion,
+		int workflowDefinitionVersion,
 		com.liferay.portal.workflow.kaleo.forms.util.TaskFormPairs taskFormPairs,
 		com.liferay.portal.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
