+++ portlets/kaleo-forms-portlet/docroot/WEB-INF/src/com/liferay/portal/workflow/kaleo/forms/service/base/KaleoProcessServiceClpInvoker.java
@@ -39,7 +39,7 @@ public class KaleoProcessServiceClpInvoker {
 
 		_methodParameterTypes54 = new String[] {
 				"long", "long", "java.util.Map", "java.util.Map", "long",
-				"java.lang.String", "long",
+				"java.lang.String", "int",
 				"com.liferay.portal.workflow.kaleo.forms.util.TaskFormPairs",
 				"com.liferay.portal.service.ServiceContext"
 			};
@@ -67,7 +67,7 @@ public class KaleoProcessServiceClpInvoker {
 
 		_methodParameterTypes59 = new String[] {
 				"long", "long", "java.util.Map", "java.util.Map", "long",
-				"java.lang.String", "long",
+				"java.lang.String", "int",
 				"com.liferay.portal.workflow.kaleo.forms.util.TaskFormPairs",
 				"com.liferay.portal.service.ServiceContext"
 			};
@@ -95,7 +95,7 @@ public class KaleoProcessServiceClpInvoker {
 				(java.util.Map<java.util.Locale, java.lang.String>)arguments[3],
 				((Long)arguments[4]).longValue(),
 				(java.lang.String)arguments[5],
-				((Long)arguments[6]).longValue(),
+				((Integer)arguments[6]).intValue(),
 				(com.liferay.portal.workflow.kaleo.forms.util.TaskFormPairs)arguments[7],
 				(com.liferay.portal.service.ServiceContext)arguments[8]);
 		}
@@ -131,7 +131,7 @@ public class KaleoProcessServiceClpInvoker {
 				(java.util.Map<java.util.Locale, java.lang.String>)arguments[3],
 				((Long)arguments[4]).longValue(),
 				(java.lang.String)arguments[5],
-				((Long)arguments[6]).longValue(),
+				((Integer)arguments[6]).intValue(),
 				(com.liferay.portal.workflow.kaleo.forms.util.TaskFormPairs)arguments[7],
 				(com.liferay.portal.service.ServiceContext)arguments[8]);
 		}
